% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{apply_endring_events}
\alias{apply_endring_events}
\title{Apply Endring (change) events to state}
\usage{
apply_endring_events(state, paat_state, endring_events, type)
}
\description{
Apply Endring (change) events to state
}
\keyword{internal}
