% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{apply_ny_events}
\alias{apply_ny_events}
\title{Apply Ny (new entity) events to state}
\usage{
apply_ny_events(state, ny_events, type)
}
\description{
Apply Ny (new entity) events to state
}
\keyword{internal}
