% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{apply_paategning_patches}
\alias{apply_paategning_patches}
\title{Apply påtegning-specific patches to the annotations state}
\usage{
apply_paategning_patches(paat_state, org_nr, patches, timestamp, update_id)
}
\description{
Apply påtegning-specific patches to the annotations state
}
\keyword{internal}
