% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replay.R
\name{apply_patch_to_row}
\alias{apply_patch_to_row}
\title{Apply a patch tibble to a new row}
\usage{
apply_patch_to_row(row, patch, valid_cols)
}
\description{
Apply a patch tibble to a new row
}
\keyword{internal}
