% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{apply_slett_events}
\alias{apply_slett_events}
\title{Apply Sletting/Fjernet events to state}
\usage{
apply_slett_events(state, paat_state, slett_events, type)
}
\description{
Apply Sletting/Fjernet events to state
}
\keyword{internal}
