% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{brreg_error_body}
\alias{brreg_error_body}
\title{Extract error messages from brreg API error responses}
\usage{
brreg_error_body(resp)
}
\arguments{
\item{resp}{An httr2 response object.}
}
\description{
Extract error messages from brreg API error responses
}
\keyword{internal}
