% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{brreg_req}
\alias{brreg_req}
\title{Build an httr2 request to a brreg API service}
\usage{
brreg_req(path, query = list(), service = c("enhetsregisteret", "fullmakt"))
}
\arguments{
\item{path}{URL path appended to the base URL.}

\item{query}{Named list of query parameters.}

\item{service}{One of \code{"enhetsregisteret"} or \code{"fullmakt"}.}
}
\description{
Build an httr2 request to a brreg API service
}
\keyword{internal}
