% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{brreg_updates_roller}
\alias{brreg_updates_roller}
\title{Fetch roller updates (CloudEvents format)}
\usage{
brreg_updates_roller(since, size)
}
\description{
Fetch roller updates (CloudEvents format)
}
\keyword{internal}
