% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{changelog_dir}
\alias{changelog_dir}
\title{Path to the changelog directory}
\usage{
changelog_dir()
}
\value{
Character path or URI.
}
\description{
Returns \verb{state_dir()/changelog/}. Changelog entries are
Hive-partitioned by \code{sync_date}.
}
\keyword{internal}
