% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet-utils.R
\name{check_parquet_available}
\alias{check_parquet_available}
\title{Abort if no parquet backend available}
\usage{
check_parquet_available()
}
\description{
Abort if no parquet backend available
}
\keyword{internal}
