% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud-io.R
\name{cloud-io}
\alias{cloud-io}
\title{Cloud and local storage abstraction}
\description{
Internal helpers that let tidybrreg I/O functions work transparently
on local file paths, GCS (\verb{gs://}) URIs, and S3 (\verb{s3://}) URIs.
Cloud storage requires the \pkg{arrow} package compiled with
GCS and/or S3 support (\code{arrow::arrow_with_gcs()},
\code{arrow::arrow_with_s3()}).
}
\keyword{internal}
