% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{compute_entries_from_bulk}
\alias{compute_entries_from_bulk}
\title{Extract daily entry counts from bulk registration_date}
\usage{
compute_entries_from_bulk(data, by)
}
\description{
Extract daily entry counts from bulk registration_date
}
\keyword{internal}
