% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.brregEnv}
\alias{.brregEnv}
\title{Session-level dictionary cache (eurostat pattern)}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.brregEnv
}
\description{
Session-level dictionary cache (eurostat pattern)
}
\keyword{internal}
