% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{empty_roller_state}
\alias{empty_roller_state}
\title{Empty roller state tibble matching \code{\link[=brreg_roles]{brreg_roles()}} schema}
\usage{
empty_roller_state()
}
\description{
Empty roller state tibble matching \code{\link[=brreg_roles]{brreg_roles()}} schema
}
\keyword{internal}
