% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{enrich_cdc}
\alias{enrich_cdc}
\title{Enrich CDC events with entity attributes from bulk data}
\usage{
enrich_cdc(updates, bulk, by)
}
\description{
Enrich CDC events with entity attributes from bulk data
}
\keyword{internal}
