% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{flatten_cell}
\alias{flatten_cell}
\title{Flatten a single list-column cell to an atomic character value}
\usage{
flatten_cell(x)
}
\description{
Flatten a single list-column cell to an atomic character value
}
\keyword{internal}
