% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{flatten_json}
\alias{flatten_json}
\title{Recursively flatten nested JSON into dot-notation keys}
\usage{
flatten_json(x, prefix = "")
}
\description{
Recursively flatten nested JSON into dot-notation keys
}
\keyword{internal}
