% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{flatten_value_into}
\alias{flatten_value_into}
\title{Recursively flatten a patch value, calling emit() for each leaf}
\usage{
flatten_value_into(op, path_prefix, value, emit)
}
\description{
Recursively flatten a patch value, calling emit() for each leaf
}
\keyword{internal}
