% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{flows_from_changelog}
\alias{flows_from_changelog}
\title{Compute flows from the persistent changelog}
\usage{
flows_from_changelog(
  by = c("nace_1", "municipality_code"),
  from = NULL,
  to = NULL
)
}
\description{
Compute flows from the persistent changelog
}
\keyword{internal}
