% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{generate_quarter_targets}
\alias{generate_quarter_targets}
\title{Generate quarter-end target dates}
\usage{
generate_quarter_targets(from, to)
}
\description{
Generate quarter-end target dates
}
\keyword{internal}
