% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{has_state}
\alias{has_state}
\title{Check whether state exists for a given type}
\usage{
has_state(type)
}
\arguments{
\item{type}{Registry type string.}
}
\value{
Logical.
}
\description{
Check whether state exists for a given type
}
\keyword{internal}
