% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label_vec}
\alias{label_vec}
\title{Label a character vector using a named dictionary}
\usage{
label_vec(x, dic, lang = "en")
}
\description{
Label a character vector using a named dictionary
}
\keyword{internal}
