% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{paginate_cdc}
\alias{paginate_cdc}
\title{Paginate through CDC events from cursor position}
\usage{
paginate_cdc(type, from_id, size = 10000L, verbose = TRUE)
}
\description{
Paginate through CDC events from cursor position
}
\keyword{internal}
