% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_entities}
\alias{parse_entities}
\title{Parse a list of raw entities into a combined tibble}
\usage{
parse_entities(raw_list)
}
\description{
Parse a list of raw entities into a combined tibble
}
\keyword{internal}
