% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{parse_updates_page}
\alias{parse_updates_page}
\title{Parse a page of raw CDC update objects into a tibble}
\usage{
parse_updates_page(raw_updates, include_changes = FALSE)
}
\arguments{
\item{raw_updates}{List of update objects from the API.}

\item{include_changes}{Logical. Parse and attach field changes.}
}
\description{
Parse a page of raw CDC update objects into a tibble
}
\keyword{internal}
