% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet-utils.R
\name{read_parquet_safe}
\alias{read_parquet_safe}
\title{Read a parquet file to tibble}
\usage{
read_parquet_safe(path)
}
\arguments{
\item{path}{Path to parquet file.}
}
\value{
A tibble.
}
\description{
Read a parquet file to tibble
}
\keyword{internal}
