% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{series_flow}
\alias{series_flow}
\title{Compute entry/exit flow series from consecutive snapshot diffs}
\usage{
series_flow(metric, by, frequency, from, to, type, label)
}
\description{
Compute entry/exit flow series from consecutive snapshot diffs
}
\keyword{internal}
