% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{to_snake}
\alias{to_snake}
\title{Convert a string from camelCase or dot.notation to snake_case}
\usage{
to_snake(x)
}
\arguments{
\item{x}{Character vector.}
}
\description{
Convert a string from camelCase or dot.notation to snake_case
}
\keyword{internal}
