% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{write_manifest_entry}
\alias{write_manifest_entry}
\title{Append an entry to the manifest}
\usage{
write_manifest_entry(entry)
}
\arguments{
\item{entry}{Named list with manifest fields.}
}
\description{
Append an entry to the manifest
}
\keyword{internal}
