% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{write_state}
\alias{write_state}
\title{Write state atomically and update session cache}
\usage{
write_state(df, type)
}
\arguments{
\item{df}{A tibble.}

\item{type}{Registry type.}
}
\description{
Write state atomically and update session cache
}
\keyword{internal}
