% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{brreg_req}
\alias{brreg_req}
\title{Build an httr2 request to the brreg API}
\usage{
brreg_req(path, query = list())
}
\arguments{
\item{path}{URL path appended to the base URL.}

\item{query}{Named list of query parameters.}
}
\description{
Build an httr2 request to the brreg API
}
\keyword{internal}
