% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{extract_paategninger}
\alias{extract_paategninger}
\title{Extract påtegninger from bulk entity data into separate table}
\usage{
extract_paategninger(entities)
}
\description{
Extract påtegninger from bulk entity data into separate table
}
\keyword{internal}
